feat: refactor Explore page structure and state management with new components
