+++ widgets/explore/ui/index.ts
@@ -1,2 +1,4 @@
 export { ExploreTabs } from './ExploreTabs';
 export { FloatingCreateButton } from './FloatingCreateButton';
+export { ExplorePageWidget } from './ExplorePageWidget';
+export { ExploreContent } from './ExploreContent';
